+++ topiary-cli/tests/sample-tester.rs
@@ -44,7 +44,6 @@ macro_rules! lang_test {
 mod test_fmt {
     use super::*;
 
-    #[allow(unused)]
     fn fmt_input(lang: &str) {
         let file = format!("{lang}.{}", get_file_extension(lang));
         let input = PathBuf::from(format!("tests/samples/input/{file}"));
@@ -145,7 +144,6 @@ mod test_fmt {
 mod test_check {
     use super::*;
 
-    #[allow(unused)]
     fn check_input(lang: &str) {
         let file = format!("{lang}.{}", get_file_extension(lang));
         let input = PathBuf::from(format!("tests/samples/input/{file}"));
@@ -199,7 +197,6 @@ mod test_check {
 mod test_coverage {
     use super::*;
 
-    #[allow(unused)]
     fn coverage_input(lang: &str) {
         let file = format!("{lang}.{}", get_file_extension(lang));
         let input = PathBuf::from(format!("tests/samples/input/{file}"));